- Replace use of memcpy for write and read pixel macros because float4 is neither plain data nor packed. Anyways a few conditionals with assignments is better optimized by compiler. - Fix scaling operations issue. Float coordinates were being casted to int even for bilinear sampling. - Fix saved files being not compatible with previous versions due to removal of tree chunksize field. chunksize field is added again just for compatibility.

+++ source/blender/compositor/buffering/COM_Pixels.h
@@ -88,9 +88,9 @@ typedef struct PixelsImg {
   /* whole rect row bytes*/
   const size_t row_bytes;
   /* buffer row elements  */
-  const int brow_elems;
+  const size_t brow_elems;
   /* buffer row channels (buffer row elems * elem channels)*/
-  const int brow_chs;
+  const size_t brow_chs;
   /* buffer row bytes (buffer row elems * elem channels)*/
   const size_t brow_bytes;
